requirements.txt
numpy==1.26.0
opencv_python==4.10.0.84
Pillow==10.4.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
pytest==8.3.3
Requests==2.32.3
setuptools==72.2.0
tensorflow==2.17.0
tensorflow_hub==0.16.1
tensorflow_intel==2.17.0
